Fix compression level extraction and remove redundant Call APIs

+++ src/utilities.cpp
@@ -42,7 +42,7 @@ bool sexp_to_bool(SEXP value) {
 }
 
 int sexp_to_int(SEXP value) {
-    return (int) *INTEGER(value);
+    return (int) *REAL(value);
 }
 
 std::string sexp_to_string(SEXP value) {
